mb/intel/harcuvar: Fix board_id() return type
The weak definition of board_id() in coreboot_table.c returns a
uint32_t, so update this function to match. This fixes a compiler error
when using LTO.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/39015
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/intel/harcuvar/harcuvar_boardid.h
@@ -8,6 +8,6 @@
 
 #define BoardIdHarcuvar 0x52
 
-uint8_t board_id(void);
+uint32_t board_id(void);
 
 #endif /* MAINBOARD_BOARD_H */
